+++ R/update_database.R
@@ -28,8 +28,8 @@
 #'
 #' @examples
 #'
-#' \donttest{
-#' update_database()
+#' \dontrun{
+#' update_database(dataset_dir = tempdir())
 #' }
 #'
 update_database <- function(